chore: split tests in multiple testsuites
