+++ sonar-project.properties
@@ -18,5 +18,5 @@ sonar.links.homepage=https://www.officelife.io/
 sonar.links.scm=https://github.com/officelifehq/officelife
 sonar.links.issue=https://github.com/officelifehq/officelife/issues
 
-sonar.php.tests.reportPath=./results/sqlite/results.xml
+sonar.php.tests.reportPath=./results/sqlite/results*.xml
 #[Code coverage not yet activated] sonar.php.coverage.reportPaths=./results/sqlite/coverage.xml
